+++ modules/manager/zz_generated_konnect_index_options.go
@@ -17,6 +17,7 @@ func generatedIndexOptionsForKonnectEntities(
 		index.OptionsForAIGatewayAgent(),
 		index.OptionsForAIGatewayAuthStrategy(),
 		index.OptionsForAIGatewayCACertificate(),
+		index.OptionsForAIGatewayCertificate(),
 		index.OptionsForAIGatewayConsumer(),
 		index.OptionsForAIGatewayConsumerCredential(),
 		index.OptionsForAIGatewayConsumerGroup(),